project31

+++ src/Controller/Admin.java
@@ -65,7 +65,7 @@ public class Admin {
             if (Objects.equals(words[i], "Add")) {
                 if (Objects.equals(words[i + 1], "car")) {
                     boolean auto;
-                    int price = Integer.parseInt(words[i + 6]);
+                    double price = Double.parseDouble(words[i + 6]);
                     int capacity = Integer.parseInt(words[i + 7]);
                     double volume = Integer.parseInt(words[i + 2]);
                     if (Objects.equals(words[i + 3], "true")) {
@@ -77,13 +77,13 @@ public class Admin {
                     Admin1.getAdmin1().getProducts().add(car);
                     Admin1.getAdmin1().getVehicles().add(car);
                 } else if (Objects.equals(words[i + 1], "bikeCycle")) {
-                    int price = Integer.parseInt(words[i + 5]);
+                    double price = Double.parseDouble(words[i + 5]);
                     int capacity = Integer.parseInt(words[i + 6]);
                     bikeCycle bike = new bikeCycle(words[i + 4], price, 0, capacity, Product.category.VEHICLES, words[i + 3], bikeCycle.bikeCycle1.valueOf(words[i + 2]),"nothing");
                     Admin1.getAdmin1().getProducts().add(bike);
                     Admin1.getAdmin1().getVehicles().add(bike);
                 } else if (Objects.equals(words[i + 1], "computer")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 8]);
                     double weight = Integer.parseInt(words[i + 4]);
                     double dimension = Integer.parseInt(words[i + 5]);
@@ -92,7 +92,7 @@ public class Admin {
                     Admin1.getAdmin1().getProducts().add(computer);
                     Admin1.getAdmin1().getDigitals().add(computer);
                 } else if (Objects.equals(words[i + 1], "information")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 6]);
                     double weight = Double.parseDouble(words[i + 4]);
                     double dimension = Double.parseDouble(words[i + 5]);
@@ -100,7 +100,7 @@ public class Admin {
                     Admin1.getAdmin1().getProducts().add(informationStorage);
                     Admin1.getAdmin1().getDigitals().add(informationStorage);
                 } else if (Objects.equals(words[i + 1], "flash")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 7]);
                     double weight = Double.parseDouble(words[i + 4]);
                     double dimension = Double.parseDouble(words[i + 5]);
@@ -108,7 +108,7 @@ public class Admin {
                     Admin1.getAdmin1().getProducts().add(flash);
                     Admin1.getAdmin1().getDigitals().add(flash);
                 } else if (Objects.equals(words[i + 1], "SSD")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 9]);
                     double weight = Double.parseDouble(words[i + 4]);
                     double dimension =Double.parseDouble(words[i + 5]);
@@ -118,7 +118,7 @@ public class Admin {
                     Admin1.getAdmin1().getProducts().add(ssd);
                     Admin1.getAdmin1().getDigitals().add(ssd);
                 } else if (Objects.equals(words[i + 1], "pen")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 6]);
                     Pen pen = new Pen(words[i + 2], price, 0, capacity, Product.category.STATIONARY, words[i + 4], words[i + 5],"nothing");
                     Admin1.getAdmin1().getProducts().add(pen);
@@ -126,20 +126,20 @@ public class Admin {
                     for (int u=0;u<Admin1.getAdmin1().getStationaries().get(u).getPens().size();u++)
                     { Admin1.getAdmin1().getStationaries().get(u).getPens().add(pen);}
                 } else if (Objects.equals(words[i + 1], "pencil")) {
-                    int price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 6]);
                     Pencil pencil = new Pencil(words[i + 2], price, 0, capacity, Product.category.STATIONARY, words[i + 4], Pencil.pencil.valueOf(words[i + 5]),"nothing");
                     Admin1.getAdmin1().getProducts().add(pencil);
                     Admin1.getAdmin1().getStationaries().add(pencil);
                 } else if (Objects.equals(words[i + 1], "notebook")) {
-                    long price = Long.parseLong(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 7]);
                     int page = Integer.parseInt(words[i + 6]);
                     noteBook noteBook = new noteBook(words[i + 2], price, 0, capacity, Product.category.STATIONARY, words[i + 4], words[i + 5], page,"nothing");
                     Admin1.getAdmin1().getProducts().add(noteBook);
                     Admin1.getAdmin1().getStationaries().add(noteBook);
                 } else if (Objects.equals(words[i + 1], "food")) {
-                    long price = Integer.parseInt(words[i + 3]);
+                    double price = Double.parseDouble(words[i + 3]);
                     int capacity = Integer.parseInt(words[i + 6]);
                     Food food = new Food(words[i + 2], price, 0, capacity, Product.category.FOOD, words[i + 4], words[i + 5],"nothing");
                     Admin1.getAdmin1().getProducts().add(food);
