+++ src/Model/Food.java
@@ -4,7 +4,7 @@ package Model;
     private String dateOfManufacture;
     private String expirationDate;
 
-    public Food(String productName,long productPrice,double averageScore,int capacity,category product,String dateOfManufacture, String expirationDate,String comment) {
+    public Food(String productName,double productPrice,double averageScore,int capacity,category product,String dateOfManufacture, String expirationDate,String comment) {
         super(productName,productPrice, averageScore,capacity,product,comment);
         this.dateOfManufacture = dateOfManufacture;
         this.expirationDate = expirationDate;
