+++ src/Model/Pen.java
@@ -3,7 +3,7 @@ package Model;
 public class Pen extends Stationary {
     private String color;
 
-   public  Pen(String productName,long productPrice,double averageScore,int capacity,category product,String country, String color,String comment) {
+   public  Pen(String productName,double productPrice,double averageScore,int capacity,category product,String country, String color,String comment) {
         super(productName,productPrice, averageScore,capacity,product,country,comment);
         this.color = color;
     }
