+++ src/Model/Product.java
@@ -30,7 +30,7 @@ public class Product {
     private int opinionNumber=0;
     private long productId;
     private int capacity;
-    private long productPrice;
+    private double productPrice;
     private String productName;
     private double averageScore;
      private String comment;
@@ -40,7 +40,7 @@ public class Product {
     {
         return opinions;
     }
-    public Product(String productName,long productPrice,double averageScore ,int capacity,category product,String comment)
+    public Product(String productName,double productPrice,double averageScore ,int capacity,category product,String comment)
     {
        this.productPrice=productPrice;
        this.averageScore=averageScore;
@@ -53,7 +53,7 @@ public class Product {
     {
         return this.productName;
     }
-    public long getProductPrice()
+    public double getProductPrice()
     {
         return this.productPrice;
     }
@@ -77,7 +77,7 @@ public class Product {
     {
         return this.capacity=capacity;
     }
-    public long  setProductprice(long price)
+    public double  setProductprice(double price)
     {
         return this.productPrice=price;
     }
