+++ src/Model/Vehicle.java
@@ -3,7 +3,7 @@ package Model;
 abstract public class Vehicle extends Product
 {
     private String company;
-    Vehicle(String productName,long productPrice,double averageScore,int capacity,category product,String company,String comment) {
+    Vehicle(String productName,double productPrice,double averageScore,int capacity,category product,String company,String comment) {
         super(productName,productPrice, averageScore,capacity,product,comment);
         this.company = company;
     }
