+++ src/Model/informationStorage.java
@@ -3,7 +3,7 @@ package Model;
 public class informationStorage extends Digital {
     private int capacity1;
 
-    public informationStorage(String productName,long productPrice,double averageScore,int capacity,category product,double weight, double dimension,String comment) {
+    public informationStorage(String productName,double productPrice,double averageScore,int capacity,category product,double weight, double dimension,String comment) {
         super(productName,productPrice, averageScore,capacity,product,weight,dimension,comment);
         this.capacity1 = capacity;
     }
