+++ src/Model/noteBook.java
@@ -4,7 +4,7 @@ public class noteBook extends Stationary {
     private int pageNumber;
     private String paperType;
 
-    public noteBook(String productName,long productPrice,double averageScore,int capacity,category product,String country, String paperType, int pageNumber,String comment) {
+    public noteBook(String productName,double productPrice,double averageScore,int capacity,category product,String country, String paperType, int pageNumber,String comment) {
         super(productName,productPrice,averageScore,capacity,product,country,comment);
         this.pageNumber = pageNumber;
         this.paperType = paperType;
